add_marker() is used in combination with whomapper() and bubblemapper(), adding a secondary marker located at the centroid of listed countries and territories.

add_marker() is used in combination with whomapper() and bubblemapper(), adding a secondary marker located at the centroid of listed countries and territories.

add_marker(
  iso3 = NA_character_,
  shape = 17,
  col = "red",
  size = 3,
  alpha = 1,
  projection = "robin",
  offset = 10.8
)

Arguments

iso3

A vector of ISO3 country codes.

shape

Marker shape.

col

Marker colour.

size

Marker size.

lab

Marker label.

Value

A ggplot2 function.

Author

Takuya Yamanaka, adapted from scripts of whomap developed by Philippe Glaziou.

Examples

brics <- data.frame(iso3=c('BRA','CHN','IND','RUS','ZAF'), var=1:5)
whomap(brics, legend.title='BRICS', legend.pos=c(0.14, 0.34)) +
   add_marker(c('BRA','RUS'))
#> Error in whomap(brics, legend.title = "BRICS", legend.pos = c(0.14, 0.34)): could not find function "whomap"