Bubble world maps 2025 version bubblemapper() prints a bubble world map based on the latest WHO geoJSON files It requires ggplot2, ggpattern, sf and here

Bubble world maps 2025 version bubblemapper() prints a bubble world map based on the latest WHO geoJSON files It requires ggplot2, ggpattern, sf and here

bubblemapper(
  X = data.frame(iso3 = NA, size = NA),
  projection = "eck4",
  offset = NULL,
  bubble_col = "dodgerblue",
  bubble_alpha = 0.4,
  scale_breaks,
  scale_limits,
  scale_labels,
  line_col = "black",
  line_width = 0.3,
  map_title = "",
  legend_title = "",
  water_col = "white",
  na_label = "No data",
  na_col = "white",
  disclaimer = FALSE,
  legend_pos = c(0.17, 0.42),
  zoom = "Global"
)

Source

Modified from WHO GIS (https://gis-who.hub.arcgis.com/)

Arguments

line_col

Colour of country border lines.

map_title

Map title.

legend_title

Legend title.

water_col

Colour of oceans and lakes.

na_label

Legend lable for missing values.

na_col

Colour of countries with missing values.

disclaimer

A boolean, inserts a standard WHO disclaimer.

legend_pos

A vector of two numbers, positions the legend.

df

A dataframe with two columns: 'iso3' (character or factor WHO country codes) and 'size' (numeric values to map as bubble sizes).

colours

A vector of colour values for each category in "var", excepting missing values.

low_col

First value of a gradient of colours.

high_col

Last value of a gradient of colours.

Value

A ggplot2 plot.

Author

Takuya Yamanaka, adapted from scripts of whomap developed by Philippe Glaziou.

Examples

bubblemapper(data.frame(iso3 = NA, size = NA))
#> Error in sf_column %in% names(g): Can't join `x$iso3` with `y$iso3` due to incompatible types.
#> ℹ `x$iso3` is a <character>.
#> ℹ `y$iso3` is a <logical>.